feat: add dropdown for keypair sub pages
